Weapons Initialization & Player Inventory
Base Weapon Class for weapons to derive off and have their own stats and firing abilities.

Added a player inventory for weapon handling, weapon input handling, and some slight changes.

This is to be the structure of the weapons in this project. I will be tackling the Visualization of the weapons and the firing functionality next.