Bullet Hit Detection with Zombies
- Completed bullet hit detection/collision for zombies.
How:
Checks every update for every zombie alive in the scene, and uses a simple circle-circle collision or the sum of the radii. If a bullet overlaps the a zombie, we call the hit function and the bullet deals the damage that has stored in it's class.

Forgot to mention last push:
- Reverted moving camera system to static camera system as the map is static. If the map is big, and you can move around in different places I will bring back the moving camera system.